examples/exchange_infobot/requirements.txt
numpy==1.13.3
pandas==1.0.1
python-dotenv==0.10.5
aiohttp==3.6.2
matplotlib==3.1.1
aiogram==2.6.1
